Consolidate net45 sub-folder
