+++ src/Microsoft.VisualStudio.Threading/RegistryChangeNotificationFilters.cs
@@ -4,6 +4,8 @@
 *                                                        *
 *********************************************************/
 
+#if DESKTOP
+
 namespace Microsoft.VisualStudio.Threading
 {
     using System;
@@ -45,3 +47,5 @@ namespace Microsoft.VisualStudio.Threading
         Security = 0x8
     }
 }
+
+#endif
